requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
Naked==0.1.31
pycrypto==2.6.1
PyYAML==5.1
requests==2.22.0
shellescape==3.4.1
urllib3==1.25.3
